+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [UITEN-227](https://issues.folio.org/browse/UITEN-227) View-only access to Settings > Tenant > Location setup > Libraries.
 * [UITEN-226](https://issues.folio.org/browse/UITEN-228) View-only access to Settings > Tenant > Location setup > Locations
 * [UITEN-20](https://issues.folio.org/browse/UITEN-20) Add Closed Library Date Management to Service Point for Hold Shelf Expiration
+* [UITEN-237](https://issues.folio.org/browse/UITEN-237) Update "Closed library date management for hold shelf expiration date calculation" menu
 
 ## [7.3.0](https://github.com/folio-org/ui-tenant-settings/tree/v7.3.0)(2022-10-20)
 [Full Changelog](https://github.com/folio-org/ui-tenant-settings/compare/v7.2.0...v7.3.0)